Refactor rapidAPIfetch: extract persistence logic into persistSearchResults service for improved modularity
